models/dependency: Move `add_dependencies()` function into to `publish` controller
